Add tests: convertRowCol, first move; refactoring
Had an issue where the correct row/col was updated, but incorrect row/col was reported. Added
test cases to correct this.

Testing that the correct player makes the first move using
mocked input

Testing a range of parameters for instantiating TicTacToe's with subTest

Refactored assertEqual()'s to check the turn equals a Turn instance
rather than just checking the value of the turn against a string value