Implement last_move; Fix broken tests & refactor
last_move is a dict with keys 'row' and 'col' used to save previous move.
This is updated whenever apply_move is called, and can be used to easily
display the computer's move or for future enhancements to get_winner

All variables and methods use snake_case format, this is updated in tests.

Tests also reflect choice to use full 'player', 'computer' or 'random'
instead of 'p', 'c' or 'r' when instantiating a TicTacToe class.