Add variable cell_width; Refactoring
Cell's can now be 1, 3, or 5 character's wide determined by the -b flag,
and the value's are either a light shading or a dark shading instead of
X's and O's so that the board looks nicer with a larger cell_width.

Added docstrings to many of the functions, and comments for much of the
code. Converted some variables into snake_case for consistency.

Created show_info() to display initial game dimensions, first move, etc.